Moved documentation component out of the page
